# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "znskr api",
    "description": "znskr paas api for managing apps and deployments",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/apps/{id}": {
      "get": {
        "tags": [
          "apps"
        ],
        "summary": "get a single app by id",
        "operationId": "get_app",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "app id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "app details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "apps"
        ],
        "summary": "update an app",
        "operationId": "update_app",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "app id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAppRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "app updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "domain conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "apps"
        ],
        "summary": "delete an app",
        "operationId": "delete_app",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "app id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "app deleted"
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/containers": {
      "get": {
        "tags": [
          "containers"
        ],
        "operationId": "list_containers",
        "responses": {
          "200": {
            "description": "list of containers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ContainerListItem"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/containers/{id}/files/download": {
      "get": {
        "tags": [
          "containers"
        ],
        "operationId": "download_volume_entry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "container id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mount",
            "in": "query",
            "description": "mount point",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "relative path within mount",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "file download"
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/containers/{id}/files/mkdir": {
      "post": {
        "tags": [
          "containers"
        ],
        "operationId": "create_volume_directory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "container id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mount",
            "in": "query",
            "description": "mount point",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "description": "directory path to create",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "directory created"
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppResponse": {
        "type": "object",
        "description": "app response",
        "required": [
          "id",
          "name",
          "github_url",
          "branch",
          "domains",
          "port",
          "env_vars",
          "services",
          "rollout_strategy",
          "created_at"
        ],
        "properties": {
          "branch": {
            "type": "string",
            "description": "branch being deployed"
          },
          "created_at": {
            "type": "string",
            "description": "creation timestamp"
          },
          "domain": {
            "type": [
              "string",
              "null"
            ],
            "description": "custom domain"
          },
          "domains": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "custom domains"
          },
          "env_vars": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnvVarResponse"
            },
            "description": "environment variables"
          },
          "github_url": {
            "type": "string",
            "description": "github repository url"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "unique app id"
          },
          "name": {
            "type": "string",
            "description": "app name"
          },
          "port": {
            "type": "integer",
            "format": "int32",
            "description": "app port (deprecated)",
            "minimum": 0
          },
          "rollout_strategy": {
            "type": "string",
            "description": "rollout strategy"
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ServiceResponse"
            },
            "description": "container services"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "error response",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "error message"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}